Force Devise to include :locale in generated URLs
On unauthenticated access, Devise tries to redirect the user to the
new_user_session_url, but doesn't pass in the required :locale parameter.
To ensure that this redirect works, we have to globally tell the Rails router
to include :locale in every generated path.

For admin controllers we don't need this, and so override the global setting
to not include the :locale.