More robust logic for Item display name
The previous logic would just show the raw JSON value of the primary (or first)
field. If the field was anything other than a simple scalar value, this would
be problematic.

To improve, this commit adds a `item_display_name` view helper that renders the
primary "human readable" field. We now use this helper throughout the app
where we need to display an item in a drop-down menu, in navigation, etc.