Add JSON support at the DataStore level
Rewrite the JSON support, so that rather than being a mixin at the field
level, it is done at the storage level. This means that every attribute
automatically gains a `_json` version of the accessor for reading and
writing data, regardless of the field being used.

When a raw JSON string is assigned to the `_json` virtual attribute, it
is parsed and the resulting object is passed to the underlying attribute
that is stored in the database.

Invalid JSON is handled gracefully, generating a validation error.

When reading a `_json` value, any Array or Hash object in the underlying
attribute is serialized into a string in JSON format. If the value is
not present, the `_json` will be `nil`. Finally, if the value is a
primitive (e.g. an Integer), it will be serialized as a JSON string like
`{ "raw_value": <<primitive>> }`.

Finally, to allow data to be migrated from old primitive formats to new
JSON representations, a Field can now provide a `transform_value` method
that can transform the original value as necessary before it gets
serialized to JSON.

All fields that were using `HasJsonRepresentation` have been migrated to
this new system.