Ensure date picker works when initial value is ""
The new JSON behavior is to return null rather than a JSON string when
there is no existing value. In the `<textarea>` that is rendered to pass
the value to the React component, the value is a blank string. This
commit ensures that a blank string is properly interpreted as an empty
object.