Use hidden input to pass data to React components
This commit standardizes how we pass data for a JSON field to React.
Rather than rendering a `<textarea>` that needs to be hidden via CSS, we
render an `<input type=hidden>` with various data attributes, like this:

```
<input data-field-uuid="_1e7d3604_57ed_4cc6_954d_9cd1b808e9f6"
       data-field-options="{&quot;format&quot;:&quot;YMD&quot;}"
       data-field-type="datetime"
       data-field-label="Published"
       data-field-required="true"
       data-field-i18n="false"
       data-field-multiple="false"
       id="item__1e7d3604_57ed_4cc6_954d_9cd1b808e9f6_json"
       type="hidden"
       value="{&quot;D&quot;: 7, &quot;M&quot;: 4, &quot;Y&quot;: 1999}"
       name="item[_1e7d3604_57ed_4cc6_954d_9cd1b808e9f6_json]">
```

The React component can then get all of the data it needs from this
hidden input, so we don't need to explicitly specify props. The only
prop passed to the component is, e.g.:

```
<div data-react-class="DateTimeInput"
     data-react-props="{&quot;input&quot;:&quot;#item__1e7d3604_57ed_4cc6_954d_9cd1b808e9f6_json&quot;}">
</div>
```

In other words, we pass the DOM `#id` of the hidden input to the React
component so that it can easily locate the input.

This system should allow for non-React components to work as well:
anything that can read the DOM can obtain what it needs from the various
data attributes of the hidden input.