<select> for UI component choices, as appropriate
Instead of displaying a text field for both the display_component and
editor_component of a field, conditionally show a UI based on the
configuration of the field type.

If no choices are allowed (i.e. in fields.json), then don't prompt the
administrator to pick a component at all. Likewise, if there is only one
component available, automatically assign it and don't let the user make
a choice.

If there are multiple component choices available, then allow the admin
to pick which one they want using a <select> list.