Display form errors for the file attachment field
The attachment_field form helper from refile/refile package don't
seems to support inline form error.

Displays manually the errors from the file field after the field.

The "is-invalid" class added to the field has currently no effect.
However, in the 5.0.0 update of bootstrap-form, the errors_on
helper will use the "invalid-feedback" class for displaying the error.
This class need to be used with "is-invalid" to work correctly.
Added this class preventively.