Fix category fields visible even when corresponding choice not selected (#575)
* Fix category fields visible even when corresponding choice not selected

When updating to Bootstrap 5, the form-group class was replaced by mb-3.
Javascript code that used this class has selector didn't work anymore.

To avoid further issues, we manually added a class instead of using the
new "mb-3" selector from Bootstrap 5. This custom class is not linked to
Bootstrap. It identifies wrapper of elements in a category that need to
be hidden or shown when the corresponding choice is selected. We chose
"form-group," as before, because the name is fitting for its intended
purpose.

* Add tests checking that field is correctly hide