Fix reference field creaiton in category
Creating a reference field within a category is complex because the
reference needs to occur between two item types. When a reference is
created in a category, it attempts to establish a reference between the
category and an item type, which leads to the bug. A category cannot be
one of the item types. However, we can't know in advance which item type
will be the 'source'—it will depend on the item type in which this
category is used.

This is a unique case that requires in-depth analysis and structural
changes.

The same happens if we decide to allow a category to be the source of
the reference. It raises a lot of questions that need to be adressed.

This commit remove the abilty to create a reference field within a
category. This specific case will be adressed if the need arise.