Fix reference field creation in category
Do not display reference label for categories when creating a reference
field.

The reference label is not relevant when the reference field is related
to a category. So we use the name of the field as label like before the
reference label was introduced.

The reference label is used to name the reverse relation. In the case of
a category, the source is a category and not an item. The reference
label is only used when the source item is displayed. Since there is no
source item to be displayed, it is irrelevant to set a reference label.