Change route pages#item name to better suit its purpose
This route is only used to retrieve JSON data used in line view. The
name and useless query string params `style` is misleading. Plus the
actual name is way to broad and make it hard to find usages.

Renaming it with a more specific name that illustrate better its
purpose.