Fix reference in category bug (#577)
* Rubocop 'Do not use instance variables in helpers.'

* Fix reference field creaiton in category

Creating a reference field within a category is complex because the
reference needs to occur between two item types. When a reference is
created in a category, it attempts to establish a reference between the
category and an item type, which leads to the bug. A category cannot be
one of the item types. However, we can't know in advance which item type
will be the 'source'—it will depend on the item type in which this
category is used.

This is a unique case that requires in-depth analysis and structural
changes.

The same happens if we decide to allow a category to be the source of
the reference. It raises a lot of questions that need to be adressed.

This commit remove the abilty to create a reference field within a
category. This specific case will be adressed if the need arise.

* Revert "Fix reference field creaiton in category"

This reverts commit bb47e4499dc31c9bba301dc6986fd42faeaef3f7.

It appears that reference field in categories was used in the past and
working. This commit [1] introduced the bug that prevent them to be
created.

Reverting this commit and fixing the bug introduced instead.

[1]: fb72572a37adca8c6557918467203e4e6e8bd707

* Fix reference field creation in category

Do not display reference label for categories when creating a reference
field.

The reference label is not relevant when the reference field is related
to a category. So we use the name of the field as label like before the
reference label was introduced.

The reference label is used to name the reverse relation. In the case of
a category, the source is a category and not an item. The reference
label is only used when the source item is displayed. Since there is no
source item to be displayed, it is irrelevant to set a reference label.

* Fix reference field in category visible even when choice not selected

The form-group class was not put on the correct html element.

Changed the wrapper_class option key by the correct one to override the
default options to avoid incorrect UI margin.

* Fix choice set validation missing field_set_type for uniqueness

* Prevent recursion between category and choice set